+++ vypocet_skore.cpp
@@ -3,7 +3,7 @@
 #include "vypocet_skore.h"
 
 // Deklarace funkce pro výpočet skóre (Vaše implementace)
-int vypocetSkore(const std::vector<int>& hody) {
+int vypocetSkore(const std::vector<double>& hody) {
     int celkoveSkore = 0;
     int indexHodu = 0; // Index pro procházení pole 'hody'
