+++ vypocet_skore.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <vector>
 
-int vypocetSkore(const std::vector<int>& hody);
+int vypocetSkore(const std::vector<double>& hody);
